* 211-fix-sign-promo-in-ctkMenuComboBoxTest1:
  Fix sign-promo warning in ctkMenuComboBoxTest1.cpp

+++ Libs/Widgets/Testing/Cpp/ctkMenuComboBoxTest1.cpp
@@ -73,7 +73,7 @@ int ctkMenuComboBoxTest1(int argc, char * argv [] )
   if (menuComboBox->toolButtonStyle() != Qt::ToolButtonIconOnly)
     {
     std::cerr << "Wrong default toolButtonStyle: "
-              << menuComboBox->toolButtonStyle()
+              << static_cast<unsigned int>(menuComboBox->toolButtonStyle())
               << std::endl;
     return EXIT_FAILURE;
     }
@@ -81,7 +81,7 @@ int ctkMenuComboBoxTest1(int argc, char * argv [] )
   if (menuComboBox->toolButtonStyle() != Qt::ToolButtonTextUnderIcon)
     {
     std::cerr << "Failed to set toolButtonStyle: "
-              << menuComboBox->toolButtonStyle()
+              << static_cast<unsigned int>(menuComboBox->toolButtonStyle())
               << std::endl;
     return EXIT_FAILURE;
     }
